+++ lib/instream/connection/config.ex
@@ -49,6 +49,14 @@ defmodule Instream.Connection.Config do
   defp maybe_fetch_deep(config, nil),  do: config
   defp maybe_fetch_deep(config, keys), do: get_in(config, keys)
 
+
+  defp maybe_fetch_system(config) when is_list(config) do
+    Enum.map config, fn
+      { k, v } -> { k, maybe_fetch_system(v) }
+      other    -> other
+    end
+  end
+
   defp maybe_fetch_system({ :system, var }), do: System.get_env(var)
   defp maybe_fetch_system(config),           do: config
 end