+++ test/helpers/env_connection.exs
@@ -1,0 +1,3 @@
+defmodule Instream.TestHelpers.EnvConnection do
+  use Instream.Connection, otp_app: :instream
+end