+++ test/test_helper.exs
@@ -2,6 +2,7 @@ Code.require_file("helpers/nil_logger.exs", __DIR__)
 
 Code.require_file("helpers/connection.exs", __DIR__)
 Code.require_file("helpers/connection_with_opts.exs", __DIR__)
+Code.require_file("helpers/env_connection.exs", __DIR__)
 Code.require_file("helpers/log_connection.exs", __DIR__)
 Code.require_file("helpers/udp_connection.exs", __DIR__)
 
@@ -20,6 +21,7 @@ alias Instream.TestHelpers
 [
   TestHelpers.Connection.child_spec,
   TestHelpers.ConnectionWithOpts.child_spec,
+  TestHelpers.EnvConnection.child_spec,
   TestHelpers.LogConnection.child_spec,
   TestHelpers.UDPConnection.child_spec,